fix: fix method update, delete notification

+++ apps/api-gateway/src/notification/notification.service.ts
@@ -6,7 +6,7 @@ import {
   NotificationWhereUnique,
   UpdateNotificationDto,
 } from '@app/common/types/notification';
-import { Inject, Injectable } from '@nestjs/common';
+import { BadRequestException, Inject, Injectable } from '@nestjs/common';
 import { MicroService } from '../grpc-client/microservice';
 import { ClientGrpc } from '@nestjs/microservices';
 import { firstValueFrom } from 'rxjs';
@@ -34,12 +34,34 @@ export class NotificationService {
     return this.notificationService.findMany(params);
   }
 
-  async update(updateDto: UpdateNotificationDto) {
+  async update(updateDto: UpdateNotificationDto, userId: string) {
+    const notification = await firstValueFrom(
+      this.notificationService.findOne({
+        id: updateDto.id,
+      }),
+    );
+
+    if (notification.userId !== userId) {
+      throw new BadRequestException('Notification not found');
+    }
     return this.notificationService.update(updateDto);
   }
 
-  async delete(where: NotificationWhereUnique) {
-    console.log('delete request' + where.id);
+  async delete(where: NotificationWhereUnique, userId: string) {
+    const notification = await firstValueFrom(
+      this.notificationService.findOne({
+        id: where.id,
+      }),
+    );
+
+    if (notification.userId !== userId) {
+      throw new BadRequestException('Notification not found');
+    }
+
     await firstValueFrom(this.notificationService.delete({ id: where.id }));
+  }
+
+  async findOne(where: NotificationWhereUnique) {
+    return this.notificationService.findOne(where);
   }
 }
